requirements.txt
# Core Dependencies
streamlit==1.31.0
pandas==2.1.4
numpy==1.26.3

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# NLP & ML
scikit-learn==1.4.0
nltk==3.8.1

# Visualization
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.1

# Utilities
python-dotenv==1.0.0
openpyxl==3.1.2

# Optional (for later)
# openai==1.10.0
# sentence-transformers==2.3.1